Deploy script is done

+++ src/main/java/com/yurets_y/payment_statistic_web/repo/StationsRepo.java
@@ -3,5 +3,5 @@ package com.yurets_y.payment_statistic_web.repo;
 import com.yurets_y.payment_statistic_web.entity.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StationsRepo extends JpaRepository<Station, Long> {
+public interface StationsRepo extends JpaRepository<Station, Integer> {
 }
